move test to dylink in core
